refactor: Differentiate between abortion type in ajax.js
The refactored code differentiates whether the abortion actually
aborted the request or if the abortion was too late and happened
once the request was complete. This can be distinguished in the
XHR's object property "stage". Look for "abortion-tried" or
"abortion-worked" within the property.

aborted-tried:  request completed and could not abort but aborting
                was tried.
aborted-worked: request successfully aborted

{master}[snippets/concepts/ajax]